pcos-surprise-discovery/requirements.txt
# Data Collection
praw==7.7.1  # Python Reddit API Wrapper
prawcore==2.4.0

# Data Processing & NLP
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1

# LLM APIs
openai==1.6.1

# Data Storage
jsonlines==4.0.0

# Analysis & Statistics
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.3.2  # For association rules mining

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
ipywidgets==8.1.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2

# Date handling
python-dateutil==2.8.2

# Data validation
pydantic==2.5.3

# PubMed API
biopython==1.83  # For PubMed queries
